feat: make wundertctl wrapper a dependency of sdk
