+++ .github/workflows/changesets.yaml
@@ -7,8 +7,6 @@ on:
       - next
     paths:
       -  '.changeset/*'
-      # allows to publish new npm packages
-      - 'packages/**/package.json'
 
 concurrency: ${{ github.workflow }}-${{ github.ref }}
