--- packages/sdk/src/wunderctl/bin/uninstall.ts
@@ -1,4 +1,0 @@
-import { uninstaller } from '@wundergraph/tools';
-import { wunderctlBinaryPath } from './../helper';
-
-export default uninstaller(wunderctlBinaryPath());
