--- packages/sdk/src/wunderctl/helper.ts
@@ -1,5 +1,0 @@
-import path from 'path';
-
-export const wunderctlBinaryPath = (): string => {
-	return path.join(__dirname, '..', '..', 'bin', 'wunderctl');
-};
