--- packages/tools/README.md
@@ -1,5 +1,0 @@
-# Tools
-
-![helper](https://img.shields.io/npm/v/@wundergraph/tools.svg)
-
-> **Warning**: This package is not intended for standalone use. No api compatibility guarantees.
