--- packages/tools/package.json
@@ -1,41 +1,0 @@
-{
-  "name": "@wundergraph/tools",
-  "version": "0.0.0",
-  "license": "Apache-2.0",
-  "description": "WunderGraph Tools",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "scripts": {
-    "build": "tsc"
-  },
-  "publishConfig": {
-    "registry": "https://registry.npmjs.org",
-    "access": "public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/wundergraph/wundergraph.git"
-  },
-  "homepage": "https://wundergraph.com",
-  "author": {
-    "name": "WunderGraph Maintainers",
-    "email": "info@wundergraph.com"
-  },
-  "bugs": {
-    "url": "https://github.com/wundergraph/wundergraph/issues"
-  },
-  "devDependencies": {
-    "@types/tar": "^4.0.4",
-    "@types/debug": "^4.1.7",
-    "@types/node": "^14.18.16",
-    "@types/rimraf": "^3.0.0",
-    "typescript": "^4.7.2"
-  },
-  "dependencies": {
-    "axios": "^0.26.1",
-    "debug": "^4.3.4",
-    "rimraf": "^3.0.2",
-    "tar": "^6.1.11"
-  },
-  "sideEffects": false
-}
