--- packages/tools/src/index.ts
@@ -1,3 +1,0 @@
-export { installer } from './wunderctl/installer';
-export { uninstaller } from './wunderctl/uninstaller';
-export { wunderctlName, downloadURL } from './wunderctl/binarypath';
