--- packages/tools/src/wunderctl/logger.ts
@@ -1,5 +1,0 @@
-import debug from 'debug';
-
-debug.enable('helper:wunderctl:error:*');
-
-export const logger = debug('helper:wunderctl');
