--- packages/tools/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "target": "ES2019",
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "declarationDir": "./dist",
-    "jsx": "react",
-    "outDir": "./dist",
-    "typeRoots": ["./node_modules/@types", "./src/typedefinitions"]
-  },
-  "include": ["src/**/*"],
-  "exclude": ["node_modules", "dist"]
-}
