+++ packages/wunderctl/.gitignore
@@ -1,1 +1,1 @@
-dist/wunderctl
+bin/wunderctl
