+++ packages/wunderctl/package.json
@@ -3,15 +3,17 @@
   "version": "0.90.23",
   "license": "Apache-2.0",
   "description": "WunderGraph Command Line Interface",
-  "main": "run.js",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "files": ["dist", "bin"],
   "bin": {
-    "wunderctl": "run.js"
+    "wunderctl": "./dist/bin/run.js"
   },
   "scripts": {
     "build": "tsc",
-    "devinstall": "node dist/install.js",
-    "postinstall": "node dist/install.js",
-    "postuninstall": "node dist/uninstall.js"
+    "devinstall": "node dist/bin/install.js",
+    "postinstall": "node dist/bin/install.js",
+    "postuninstall": "node dist/bin/uninstall.js"
   },
   "publishConfig": {
     "registry": "https://registry.npmjs.org",
@@ -30,15 +32,18 @@
     "url": "https://github.com/wundergraph/wundergraph/issues"
   },
   "devDependencies": {
-    "@wundergraph/tools": "workspace:0.0.0",
     "@types/debug": "^4.1.7",
     "@types/node": "^14.18.16",
     "@types/rimraf": "^3.0.0",
+    "@types/tar": "^4.0.4",
     "typescript": "^4.7.2"
   },
   "dependencies": {
+    "axios": "^0.26.1",
     "debug": "^4.3.4",
-    "execa": "5.1.1"
+    "execa": "5.1.1",
+    "rimraf": "^3.0.2",
+    "tar": "^6.1.11"
   },
   "sideEffects": false
 }
