--- packages/wunderctl/run.js
@@ -1,3 +1,0 @@
-#!/usr/bin/env node
-
-require('./dist/run').default();
