+++ packages/wunderctl/src/index.ts
@@ -1,0 +1,4 @@
+export {installer} from './installer';
+export {uninstaller} from './uninstaller';
+export {exec} from './exec';
+export {wunderctlName, downloadURL, wunderctlDir, wunderctlBinaryPath} from './binarypath';
