+++ packages/wunderctl/src/logger.ts
@@ -1,5 +1,5 @@
 import debug from 'debug';
 
-debug.enable('wunderctl:error:*');
+debug.enable('helper:wunderctl:error:*');
 
-export const logger = debug('wunderctl');
+export const logger = debug('helper:wunderctl');
