+++ packages/wunderctl/src/uninstaller.ts
